requirements.txt
# --- Core Data Science & Engine ---
# Se fijan versiones para garantizar que el modelo .pkl sea compatible
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn>=1.8.0
cloudpickle==3.0.0
xgboost==2.0.3
imblearn==0.0
numba==0.58.1

# --- MLOps & Tracking (Sincronizado con tu meta.yaml) ---
mlflow==2.10.2
pyyaml==6.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-proto==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-fastapi-instrumentator==6.1.0

# --- API & Deployment ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
joblib==1.3.2

# --- UI & Visualización ---
# Nota: streamlit-shap a veces entra en conflicto con shap puro, 
# se recomienda usar solo shap con st.components.v1
shap==0.44.0
streamlit==1.31.0
plotly==5.18.0
cachetools==5.3.2
requests==2.31.0

# --- Calidad de Código & Tipado (DevOps) ---
# Estas versiones aseguran que tu Mypy y Flake8 no den falsos positivos
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pytest==8.0.0
pytest-asyncio==0.23.5
types-PyYAML==6.0.12.12
types-requests==2.31.0.20240125
pandas-stubs==2.1.4.231227